+++ smart-actions/src/test/java/org/smartdata/actions/hdfs/TestSetStoragePolicyAction.java
@@ -17,20 +17,11 @@
  */
 package org.smartdata.actions.hdfs;
 
-import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.StorageType;
-import org.apache.hadoop.hdfs.DFSClient;
-import org.apache.hadoop.hdfs.DFSConfigKeys;
-import org.apache.hadoop.hdfs.DistributedFileSystem;
-import org.apache.hadoop.hdfs.MiniDFSCluster;
 import org.apache.hadoop.hdfs.protocol.HdfsFileStatus;
 import org.apache.hadoop.util.StringUtils;
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.Test;
-import org.smartdata.SmartContext;
 import org.smartdata.actions.ActionStatus;
-import org.smartdata.conf.SmartConf;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -38,7 +29,7 @@ import java.io.OutputStream;
 /**
  * Test for SetStoragePolicyAction.
  */
-public class TestSetStoragePolicyAction {
+public class TestSetStoragePolicyAction extends ActionMiniCluster {
   private static final byte MEMORY_STORAGE_POLICY_ID = 15;
   private static final byte ALLSSD_STORAGE_POLICY_ID = 12;
   private static final byte ONESSD_STORAGE_POLICY_ID = 10;
@@ -46,63 +37,30 @@ public class TestSetStoragePolicyAction {
   private static final byte WARM_STORAGE_POLICY_ID = 5;
   private static final byte COLD_STORAGE_POLICY_ID = 2;
 
-  private static final int DEFAULT_BLOCK_SIZE = 50;
-  MiniDFSCluster cluster;
-  DistributedFileSystem dfs;
-  DFSClient dfsClient;
-  SmartContext smartContext;
-
-  @Before
-  public void init() throws Exception {
-    SmartConf conf = new SmartConf();
-    initConf(conf);
-    cluster = new MiniDFSCluster.Builder(conf)
-            .numDataNodes(3)
-            .storagesPerDatanode(4)
-            .storageTypes(new StorageType[]{StorageType.DISK, StorageType.ARCHIVE,
-                    StorageType.SSD, StorageType.RAM_DISK})
-            .build();
-    cluster.waitActive();
-    dfsClient = cluster.getFileSystem().getClient();
-    smartContext = new SmartContext(conf);
-  }
-
-  static void initConf(Configuration conf) {
-    conf.setLong(DFSConfigKeys.DFS_BLOCK_SIZE_KEY, DEFAULT_BLOCK_SIZE);
-    conf.setInt(DFSConfigKeys.DFS_BYTES_PER_CHECKSUM_KEY, DEFAULT_BLOCK_SIZE);
-    conf.setLong(DFSConfigKeys.DFS_HEARTBEAT_INTERVAL_KEY, 1L);
-    conf.setLong(DFSConfigKeys.DFS_NAMENODE_REPLICATION_INTERVAL_KEY, 1L);
-  }
-
   @Test
   public void testDifferentPolicies() throws IOException {
-    try {
-      final String file = "/testStoragePolicy/file";
-      dfsClient.mkdirs("/testStoragePolicy");
+    final String file = "/testStoragePolicy/file";
+    dfsClient.mkdirs("/testStoragePolicy");
 
-      // write to HDFS
-      final OutputStream out = dfsClient.create(file, true);
-      byte[] content = "Hello".getBytes();
-      out.write(content);
-      out.close();
+    // write to HDFS
+    final OutputStream out = dfsClient.create(file, true);
+    byte[] content = "Hello".getBytes();
+    out.write(content);
+    out.close();
 
-      byte policy = setStoragePolicy(file, "ALL_SSD");
-      Assert.assertEquals(ALLSSD_STORAGE_POLICY_ID, policy);
+    byte policy = setStoragePolicy(file, "ALL_SSD");
+    Assert.assertEquals(ALLSSD_STORAGE_POLICY_ID, policy);
+    policy = setStoragePolicy(file, "COLD");
+    Assert.assertEquals(COLD_STORAGE_POLICY_ID, policy);
 
-      policy = setStoragePolicy(file, "COLD");
-      Assert.assertEquals(COLD_STORAGE_POLICY_ID, policy);
+    policy = setStoragePolicy(file, "ONE_SSD");
+    Assert.assertEquals(ONESSD_STORAGE_POLICY_ID, policy);
 
-      policy = setStoragePolicy(file, "ONE_SSD");
-      Assert.assertEquals(ONESSD_STORAGE_POLICY_ID, policy);
+    policy = setStoragePolicy(file, "HOT");
+    Assert.assertEquals(HOT_STORAGE_POLICY_ID, policy);
 
-      policy = setStoragePolicy(file, "HOT");
-      Assert.assertEquals(HOT_STORAGE_POLICY_ID, policy);
-
-      policy = setStoragePolicy(file, "WARM");
-      Assert.assertEquals(WARM_STORAGE_POLICY_ID, policy);
-    } finally {
-      cluster.shutdown();
-    }
+    policy = setStoragePolicy(file, "WARM");
+    Assert.assertEquals(WARM_STORAGE_POLICY_ID, policy);
   }
 
   private byte setStoragePolicy(String file, String storagePolicy)
